Add the beginnings of an interactive shell tutorial
Not sure if this is 100% a good idea, since it's probably going to be
subtely different from vanilla bash. Has readline and history support
though, so it's fairly convincing.

Also would need to do some work checking if the commands had been run
succesfully. Probably specifying a function to check if the output is
correct would do...